backend/requirements.txt
alembic==1.13.2
aniso8601==9.0.1
attrs==23.2.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
rpds-py==0.19.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
youtube-transcript-api==0.6.2
google-ai-generativelanguage==0.6.6
